Move base-props.ts from utils to common
