+++ js/renovation/ui/common/base_props.ts
