+++ js/renovation/ui/common/widget.tsx
@@ -23,7 +23,7 @@ import { combineClasses } from '../../utils/combine_classes';
 import { extend } from '../../../core/utils/extend';
 import { focusable } from '../../../ui/widget/selectors';
 import { normalizeStyleProp } from '../../../core/utils/style';
-import { BaseWidgetProps } from '../../utils/base_props';
+import { BaseWidgetProps } from './base_props';
 import { EffectReturn } from '../../utils/effect_return.d';
 import { ConfigContextValue, ConfigContext } from '../../common/config_context';
 import { ConfigProvider } from '../../common/config_provider';
