+++ js/renovation/ui/grids/data_grid/common/data_grid_props.tsx
@@ -9,7 +9,7 @@ import {
 } from '@devextreme-generator/declarations';
 import DxDataGrid from '../../../../../ui/data_grid';
 import type { Options, dxDataGridColumn, dxDataGridRowObject } from '../../../../../ui/data_grid';
-import { BaseWidgetProps } from '../../../../utils/base_props';
+import { BaseWidgetProps } from '../../../common/base_props';
 
 import type { dxFilterBuilderOptions } from '../../../../../ui/filter_builder';
 import { TPromise } from '../../../../../core/utils/deferred'; // eslint-disable-line import/named
