+++ js/renovation/ui/scroll_view/scroll_view_props.ts
@@ -6,7 +6,7 @@ import {
   ScrollableProps,
 } from './scrollable_props';
 
-import { BaseWidgetProps } from '../../utils/base_props';
+import { BaseWidgetProps } from '../common/base_props';
 
 @ComponentBindings()
 export class ScrollViewProps extends ScrollableProps {
