+++ js/renovation/ui/scroll_view/scrollable.tsx
@@ -10,7 +10,7 @@ import {
   ScrollableLocation, ScrollOffset,
 } from './types.d';
 
-import { BaseWidgetProps } from '../../utils/base_props';
+import { BaseWidgetProps } from '../common/base_props';
 import {
   ScrollableProps,
 } from './scrollable_props';
